requirements.txt
ably==2.0.9
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
distro==1.9.0
fastjsonschema==2.21.1
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
MarkupSafe==3.0.2
methoddispatch==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.3
openai==1.63.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
promptlayer==1.0.39
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==12.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
